+++ shell/platform/android/BUILD.gn
@@ -148,6 +148,7 @@ java_library("flutter_shell_java") {
     "io/flutter/plugin/platform/PlatformViewsController.java",
     "io/flutter/plugin/platform/SingleViewPresentation.java",
     "io/flutter/plugin/platform/VirtualDisplayController.java",
+    "io/flutter/util/BSDiff.java",
     "io/flutter/util/PathUtils.java",
     "io/flutter/util/Preconditions.java",
     "io/flutter/util/Predicate.java",
